работает рефреш страницы и логаут когда время токена вышло
