+++ src/Pages/MyCabinetPage/MyCabinetPage.jsx
@@ -1,19 +1,10 @@
 // import ExitButton from '../../Components/ExitButton/ExitButton';
 import HeaderPage from '../../Components/HeaderPage/HeaderPage';
 import DashboardList from '../../Components/DashboardList/DashboardList';
-import Selector from '../../Redux/Todos/todosSelectors';
-import { useDispatch, useSelector } from 'react-redux';
-import { refToken } from '../../Redux/Operations/authOperation';
-import { error } from '../../Redux/Selectors/authSelectors';
+import RefreshToken from '../../Components/RefreshToken/RefreshToken';
 
 const MyCabinetPage = () => {
-  const state = useSelector(Selector.getErrorRefToken);
-  const stateAuth = useSelector(error);
-  const dispatch = useDispatch();
-  if (state || stateAuth === 401) {
-    dispatch(refToken());
-    console.log('была ошибка 401');
-  }
+  RefreshToken();
 
   return (
     <>
