+++ src/Pages/SingUpPage/SingUpPage.jsx
@@ -2,12 +2,8 @@ import React, { useState } from 'react';
 import { NavLink } from 'react-router-dom';
 import '../LoginPage/Login.scss';
 import s from './SingUpPage.module.scss';
-import { useDispatch, useSelector } from 'react-redux';
-import { getWaiting } from '../../Redux/Selectors/authSelectors';
+import { useDispatch } from 'react-redux';
 import { registerAuth } from '../../Redux/Operations/authOperation';
-import { CSSTransition } from 'react-transition-group';
-// import Loader from 'react-loader-spinner';
-// import 'react-loader-spinner/dist/loader/css/react-spinner-loader.css';
 import bgMobile from '../../Images/bg-mobile.png';
 import bgMobile2 from '../../Images/bg-mobile2.png';
 import pic1 from '../../Images/pic1.png';
@@ -16,7 +12,6 @@ import Spinner from '../../Components/Spinner/Spinner';
 
 const SingUpPage = () => {
   const dispatch = useDispatch();
-  const wait = useSelector(getWaiting);
   const [name, setName] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
@@ -24,8 +19,6 @@ const SingUpPage = () => {
   const nameFunc = e => setName(e.target.value);
   const emailFunc = e => setEmail(e.target.value);
   const passwordFunc = e => setPassword(e.target.value);
-  // const emailFunc = e => setEmail(e.target.value);
-  // const passwordFunc = e => setPassword(e.target.value);
 
   const btnClick = e => {
     e.preventDefault();
@@ -36,21 +29,7 @@ const SingUpPage = () => {
 
   return (
     <section className={s.registr_section}>
-      <CSSTransition
-        in={wait}
-        classNames="waitingMessage"
-        timeout={250}
-        unmountOnExit
-      >
-        <Spinner />
-        {/* <Loader
-          className={s.waitingMessage}
-          type="BallTriangle"
-          color="#00BFFF"
-          height={100}
-          width={100}
-        /> */}
-      </CSSTransition>
+      <Spinner />
       <div className={s.registr_container}>
         <h1 className={s.registr_title}>Questify</h1>
         <h2 className={s.registr_caption}>Registration</h2>
